changed int to double in dictionary

+++ src/Services/ReviewService/ReviewService.Domain/Models/AverageReviewModel.cs
@@ -3,5 +3,5 @@ namespace ReviewService.Domain.Models;
 public class AverageReviewModel
 {
     public required double Average { get; set; }
-    public required Dictionary<int,int> TrackRatings { get; set; }
+    public required Dictionary<int,double> TrackRatings { get; set; }
 }
